+++ AppStore/AppStore/Presentation/Detail/ViewModel/DetailItem.swift
@@ -5,29 +5,7 @@
 ////  Created by summercat on 2023/07/12.
 ////
 
-import Foundation
 import UIKit
-
-enum DetailSection: Int, Hashable, CaseIterable {
-    case top
-    case summary
-    case screenshot
-    case description
-//    case rating
-    case releaseNote
-    case information
-}
-
-extension DetailSection {
-    func isValid(for app: App) -> Bool {
-        switch self {
-        case .releaseNote:
-            return app.releaseNotes != nil && app.releaseNotes != .init()
-        default:
-            return true
-        }
-    }
-}
 
 enum DetailItem: Hashable {
     case top(DetailTopItemModel)
